### 2017-03-11 #203 @arfon
👋 @arokem @jakevdp @danielskatz - would one of you be willing to edit this submission for JOSS?
