### 2017-04-11 #235 @whedon
[REVIEW]: Text detection in screen images with a Convolutional Neural Network
**Submitting author:** @domoritz (<a href="http://orcid.org/0000-0002-3110-1053">ROBYN DOMINIK MORITZ</a>)
**Repository:** <a href="https://github.com/domoritz/label_generator" target ="_blank">https://github.com/domoritz/label_generator</a>
**Version:** v1.1
