### 2017-05-03 #235 @ekolve
I followed all the steps in the README to install on AWS. I am using Ubuntu 14.04 and when I run this:
`python label_gen.py read-s3 escience.washington.edu.viziometrics acl_anthology/pdf/C08-1099.pdf acl_anthology`

I get the following message:
Usage:
`  main.py read-s3 S3-IN-BUCKET S3-FILE S3-OUT-BUCKET S3-PATH [--use-ramdisk] [--debug] [--dbg-image]
  main.py read FILE PATH [--debug] [--dbg-image]
  main.py (-h | --help)
  main.py --version`
