### 2017-05-03 #235 @domoritz
@ekolve Thanks for doing the review! It looks like `label_gen.py` now expects two buckets. I'm looking into it.
