### 2017-05-03 #235 @ekolve
Running this command:
` parallel --resume -j +6 --no-run-if-empty --eta --joblog /tmp/par.log python label_gen.py read-s3 escience.washington.edu.viziometrics acl_anthology/pdf/{} escience.washington.edu.viziometrics acl_anthology --dbg-image :::: acl_papers.txt`

I get several 403 Forbidden errors.  This was caused by me not having permission to the the s3 output bucket.  It would probably be a good idea to specify in the example that the user must supply their own output bucket they have write permission to.
