### 2017-05-03 #235 @ekolve
A few documentation requests:
Specify the platform - I assumed Ubuntu 14.04 because that is what PDF figures requires, but it would be good for people to see this documented here.

I had to install the following additional apt packages:
software-properties-common # to get the add-apt-repository utility
python-dev
liblapack-dev libblas-dev gfortran # scipy dependencies - installed from requirements.txt

Not necessarily, specific to your code, but I had to run:
`pip install awscli --force-reinstall --upgrade`
to resolve an issue with aws + botocore.session followed by `aws configure`
