### 2017-05-03 d1e737b
Update label generator code to latest pdfigures
