### 2017-05-03 #235 @ekolve
It would also be a good idea to either freeze your requirements.txt or at least pin the packages to a major version (e.g. scipy >=1.12, < 1.13).  That way if a package deprecates something you are depending on in a new version users will be able to install.
