### 2017-05-03 #235 @domoritz
Okay, first step was to get pdffigures working again. The api changed a bit (pdffigures now generates only a single json file) so I had to change the label generator code a bit but local generation works again now.

```
python label_gen.py read papers/1504.01101.pdf /tmp --debug                                                                                       1265ms
DEBUG:root:Run pdffigures /Users/domoritz/Developer/vision/label_generator/papers/1504.01101.pdf
DEBUG:root:Finished. Now look for the JSON and generate labels.
DEBUG:root:Found 4 figures
DEBUG:root:Render image /tmp/img/1504.01101-Figure-0.png from /Users/domoritz/Developer/vision/label_generator/papers/1504.01101.pdf
DEBUG:root:Render image /tmp/img/1504.01101-Figure-0-2x.png from /Users/domoritz/Developer/vision/label_generator/papers/1504.01101.pdf
DEBUG:root:generate label /tmp/text-masked/1504.01101-Figure-0-label.png
DEBUG:root:Render image /tmp/img/1504.01101-Figure-1.png from /Users/domoritz/Developer/vision/label_generator/papers/1504.01101.pdf
DEBUG:root:Render image /tmp/img/1504.01101-Figure-1-2x.png from /Users/domoritz/Developer/vision/label_generator/papers/1504.01101.pdf
DEBUG:root:generate label /tmp/text-masked/1504.01101-Figure-1-label.png
DEBUG:root:Render image /tmp/img/1504.01101-Figure-2.png from /Users/domoritz/Developer/vision/label_generator/papers/1504.01101.pdf
DEBUG:root:Render image /tmp/img/1504.01101-Figure-2-2x.png from /Users/domoritz/Developer/vision/label_generator/papers/1504.01101.pdf
DEBUG:root:generate label /tmp/text-masked/1504.01101-Figure-2-label.png
DEBUG:root:Render image /tmp/img/1504.01101-Figure-3.png from /Users/domoritz/Developer/vision/label_generator/papers/1504.01101.pdf
DEBUG:root:Render image /tmp/img/1504.01101-Figure-3-2x.png from /Users/domoritz/Developer/vision/label_generator/papers/1504.01101.pdf
DEBUG:root:generate label /tmp/text-masked/1504.01101-Figure-3-label.png
```

I'm looking at S3 now.
