### 2017-05-03 c7d3f92
Document versions and freeze dependencies
