### 2017-05-03 #235 @domoritz
> Specify the platform - I assumed Ubuntu 14.04 because that is what PDF figures requires, but it would be good for people to see this documented here.

Done

> It would also be a good idea to either freeze your requirements.txt or at least pin the packages to a major version (e.g. scipy >=1.12, < 1.13). That way if a package deprecates something you are depending on in a new version users will be able to install.

Done

The last remaining issue are the changes in AWS. The problem seems to be that a) I didn't document that you need to write to a bucket that you have write access to and b) that the `label_gen.py` arguments are not correct in the instructions, right?
