### 2017-05-03 a08c3ca
Fix label_gen command in example.
