### 2017-05-03 #235 @ekolve
Correct.  I am running the parallel task now and it seems to be running fine.
