### 2017-05-03 #235 @domoritz
I don't have access to an AWS machine right now but I fixed the command in the readme and everything seems to be working fine on my local machine.
