### 2017-05-03 bc97540
Note in readme that the commends are the commands I used.
