### 2017-05-03 #235 @domoritz
That's great. I fixed the command in the readme and everything seems to be working fine on my local machine. I also added a note to the readme that the commands were the ones I used to extract the figures and generate the labels that you can also just download from S3 (as described in the readme). I think the relevant part of those instructions is that you can use it to understand how you can generate the text masks for papers that I have not processed yet.

After the upgrades I did today, I think the code and docs are at a point where we can do that.
