### 2017-05-03 #235 @domoritz
Thank you for the review and the suggestions to improve the docs and code @ekolve!
