### 2017-05-04 #235 @ekolve
I am getting an error running the find_bad.py command:
`python find_bad.py read-s3 escience.washington.edu.viziometrics acl_anthology/json > anthology_bad.txt`
Run 0 of 1

ERROR:root:Not a json file acl_anthology/json/

This might be coming from boto.
