### 2017-05-04 5fc06f7
Delete json file after splitting, use python3 style print
