### 2017-05-04 d2a7eb5
Ignore the directory itself when listing files.
