### 2017-05-04 #235 @domoritz
Looks like boto has changed its behavior. I'm asking boto to enumerate all the files in a directory and they now also list the directory itself. Strange but not a big problem. Fixed in https://github.com/domoritz/label_generator/commit/d2a7eb54e93898df005a416807d65f3993b3912c
