### 2017-05-05 #235 @ekolve
I am getting an error running this command:
` python label_gen.py read testdata/paper.pdf /tmp/test --dbg-image --debug`

DEBUG:root:Run pdffigures /home/erick/label_generator/testdata/paper.pdf
DEBUG:root:Finished. Now look for the JSON and generate labels.
Traceback (most recent call last):
  File "label_gen.py", line 193, in <module>
    arguments['--dbg-image'], False)
  File "label_gen.py", line 94, in run_local
    with open(json_file) as fh:
IOError: [Errno 2] No such file or directory: '/tmp/test/json/paper.json'

I am using the latest code from master.
