### 2017-05-08 #235 @ekolve
I am trying to re-run the label generation:
`parallel --resume -j +6 --no-run-if-empty --eta --joblog /tmp/par.log python label_gen.py read-s3 escience.washington.edu.viziometrics acl_anthology/pdf/{} escience.washington.edu.viziometrics acl_anthology --dbg-image :::: acl_papers.txt
`
And I am getting this error:
  File "label_gen.py", line 190, in <module>
    arguments['--use-ramdisk'], arguments['--dbg-image'])
  File "label_gen.py", line 162, in run_s3
    files = run_local(target, dirpath, debug_image, True)
  File "label_gen.py", line 94, in run_local
    with open(json_file) as fh:
IOError: [Errno 2] No such file or directory: '/tmp/tmpyVYzU7/A83-1016.json'

over and over again.
