### 2017-05-10 #235 @domoritz
I just tested `python label_gen.py read testdata/paper.pdf /tmp/test --dbg-image --debug` and get

```
DEBUG:root:Run pdffigures /Users/domoritz/Developer/vision/label_generator/testdata/paper.pdf
DEBUG:root:Finished. Now look for the JSON and generate labels.
DEBUG:root:Found 6 figures
DEBUG:root:Render image /tmp/test/img/paper-Figure-0.png from /Users/domoritz/Developer/vision/label_generator/testdata/paper.pdf
DEBUG:root:Render image /tmp/test/img/paper-Figure-0-2x.png from /Users/domoritz/Developer/vision/label_generator/testdata/paper.pdf
DEBUG:root:generate label /tmp/test/text-masked/paper-Figure-0-label.png
DEBUG:root:Render image /tmp/test/img/paper-Figure-1.png from /Users/domoritz/Developer/vision/label_generator/testdata/paper.pdf
DEBUG:root:Render image /tmp/test/img/paper-Figure-1-2x.png from /Users/domoritz/Developer/vision/label_generator/testdata/paper.pdf
DEBUG:root:generate label /tmp/test/text-masked/paper-Figure-1-label.png
DEBUG:root:Render image /tmp/test/img/paper-Figure-2.png from /Users/domoritz/Developer/vision/label_generator/testdata/paper.pdf
DEBUG:root:Render image /tmp/test/img/paper-Figure-2-2x.png from /Users/domoritz/Developer/vision/label_generator/testdata/paper.pdf
DEBUG:root:generate label /tmp/test/text-masked/paper-Figure-2-label.png
DEBUG:root:Render image /tmp/test/img/paper-Figure-3.png from /Users/domoritz/Developer/vision/label_generator/testdata/paper.pdf
DEBUG:root:Render image /tmp/test/img/paper-Figure-3-2x.png from /Users/domoritz/Developer/vision/label_generator/testdata/paper.pdf
DEBUG:root:generate label /tmp/test/text-masked/paper-Figure-3-label.png
DEBUG:root:Render image /tmp/test/img/paper-Figure-4.png from /Users/domoritz/Developer/vision/label_generator/testdata/paper.pdf
DEBUG:root:Render image /tmp/test/img/paper-Figure-4-2x.png from /Users/domoritz/Developer/vision/label_generator/testdata/paper.pdf
DEBUG:root:generate label /tmp/test/text-masked/paper-Figure-4-label.png
DEBUG:root:Render image /tmp/test/img/paper-Figure-5.png from /Users/domoritz/Developer/vision/label_generator/testdata/paper.pdf
DEBUG:root:Render image /tmp/test/img/paper-Figure-5-2x.png from /Users/domoritz/Developer/vision/label_generator/testdata/paper.pdf
DEBUG:root:generate label /tmp/test/text-masked/paper-Figure-5-label.png
```

I suspect that you get the second error because of permission issues on S3, which I consider outside the scope of this review. When running over S3 data, the code always just calls the same method that is used for local generation (https://github.com/domoritz/label_generator/blob/master/label_gen.py#L162).
