### 2017-05-10 #235 @ekolve
Looks like I am getting an error from pdffigures:

Running this pdf: X98-1018.pdf

pdffigures/pdffigures -j /tmp/tmp70kZSJ/X98-1018 /tmp/tmp70kZSJ/X98-1018.pdf

Gives this output:
Body text appears to be encoded as graphics, skipping (use -i to parse these kinds of documents)
Running pdffigures version 1.0.6
