### 2017-05-10 #235 @domoritz
Interesting. I think this should only be a warning. Does it crash the
script?
