### 2017-05-10 #235 @ekolve
It doesn't crash the script, but pdffigures doesn't write anything out so there is no json to read which causes the code to failure a few lines later.
