### 2017-05-10 #235 @ekolve
Okay, I not sure what is going on with pdffigures, but this is what I am seeing:

`pdffigures/pdffigures -j /tmp/X98-1018 /tmp/X98-1018.pdf.pdf`

works - I get a file in /tmp/X98-1018.json

If i run:

`mkdir /tmp/foo`
`pdffigures/pdffigures -j /tmp/foo/X98-1018 /tmp/X98-1018.pdf.pdf`

no json file gets placed in /tmp/foo
