### 2017-05-10 #235 @domoritz
Yeah, I can confirm. The directory needs to exist.

```
pdffigures/pdffigures -j /tmp/json/test ./testdata/paper.pdf
```

Only works if `tmp/json` exists.
