### 2017-05-10 #235 @domoritz
Running

```
python label_gen.py read testdata/paper.pdf /tmp/test --dbg-image
```

Works even when only `/tmp` exists and it generates the labels correctly for me. I'm using pdffigures version 1.0.6.

Again, thank you so much for putting the effort into testing the label generator thoroughly. I'm at a conference so please excuse the slow replies lately.
