### 2017-05-10 #235 @domoritz
> Could you try the pdf that I was using?

Sure! Can you send me a link to the file or the S3 location?
