### 2017-05-10 #235 @domoritz
Found it: https://s3-us-west-2.amazonaws.com/escience.washington.edu.viziometrics/acl_anthology/pdf/X98-1018.pdf
