### 2017-05-10 8c665df
Pdffigures doesn’t create a json file if there are no figures. We have to check for the json file before trying to open it.
