### 2017-05-10 #235 @domoritz
Ahh, yes. That was a bug. The problem was that pdffigures doesn't create a JSON file when it couldn't extract any figures from the paper.

Fixed in https://github.com/domoritz/label_generator/commit/8c665df0ec37f249818455b7c100fdec5440605e. I just added a check for whether the JSON file exists.
