### 2017-05-19 #235 @ekolve
I am going to try and train the network - I pulled down all the data that I generated using the parallel command from s3. Now I have a directory acl_anthology with three directories inside of it: img, json, text-masked.  Which directory of images do I use to train the network?
