### 2017-05-19 #235 @domoritz
@ekolve I think the main contribution of my code (and what others are using it for) is the label generation. IMHO, you don't necessarily need to do the training because I just use darknet and these days better models exist.
