### 2017-05-19 #235 @ekolve
Okay, that sounds good.

The only remaining points I have to verify are: Community guidelines and Automated Tests.

Could you add a section to the README to satisfy community guidelines (description for how to contribute, etc.)?

For automated tests, I am able to run the example label_gen.py you have in the README, but I don't necessarily know if my output is what is expected.  Could you provide some way to verify that its correct?
