### 2017-05-20 #235 @domoritz
Thank you! I'll do that today. For intimated tests, I should probably add a
test that runs a full image generation on Travis and compares the outputs.
