### 2017-05-21 cd03f4a
Add community guidelines and basic travis test.
