+++ requirements.txt
@@ -1,9 +1,10 @@
 appdirs==1.4.3
 astroid==1.4.9
-awscli==1.11.82
+awscli==1.11.89
 backports.functools-lru-cache==1.3
 boto==2.46.1
-botocore==1.5.45
+botocore==1.5.52
+click==6.7
 colorama==0.3.7
 configparser==3.5.0
 cycler==0.10.0
@@ -12,15 +13,14 @@ docopt==0.6.2
 docutils==0.13.1
 flake8==2.5.4
 functools32==3.2.3.post2
+future==0.16.0
 futures==3.1.1
 isort==4.2.5
 jmespath==0.9.2
 lazy-object-proxy==1.2.2
-matplotlib==2.0.0
 mccabe==0.4.0
 networkx==1.11
 nose==1.3.7
-numpy==1.12.1
 olefile==0.44
 packaging==16.8
 pep8==1.7.0
@@ -40,7 +40,6 @@ PyYAML==3.12
 rsa==3.4.2
 s3transfer==0.1.10
 scikit-image==0.13.0
-scipy==0.19.0
 six==1.10.0
 subprocess32==3.2.7
 virtualenv==15.0.3
