### 2017-05-22 #235 @domoritz
I've added some guidelines and set up basic tests on travis. As a basic test, compare your output of `python label_gen.py read testdata/paper.pdf /tmp/test --dbg-image` with the content of [`testoutput`](https://github.com/domoritz/label_generator/tree/master/testoutput). I also created a new [tag 1.2](https://github.com/domoritz/label_generator/releases/tag/1.2).
