### 2017-05-22 38d0dc8
Test debug images on travis
