### 2017-06-06 #235 @ekolve
lgtm, want me to close out the issue?
