### 2017-06-06 #235 @domoritz
@ekolve Can you check the last boxes in this issue?

@jakevdp What happens next?
