### 2017-06-25 #235 @domoritz
Ping @ekolve @jakevdp
