### 2017-07-04 #235 @arfon
Thanks @domoritz.

> How will we make sure that the doi links to the code on GitHub instead of the uploaded file on Zenodo?

Your paper (and the JOSS website) link directly to GitHub. I don't know about the Zenodo linking sorry other than I think they include links to the GitHub repository in their metadata for the DOI.
