### 2017-07-04 5741869
Update paper.md
